// admin-dashboard/src/app/services/models/page-response-shipment-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ShipmentSummary } from '../models/shipment-summary';
export interface PageResponseShipmentSummary {
  content?: Array<ShipmentSummary>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}