// admin-dashboard/src/app/services/models/shipment-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ShipmentSummary {
  createdAt?: string;
  currentStatus?: 'CREATED' | 'ADMIN_REQUESTED_CHANGE' | 'CONSIGNOR_COUNTERED' | 'ADMIN_REJECTED_OFFER' | 'CONSIGNOR_REJECTED_OFFER' | 'CONSIGNOR_ACCEPTED' | 'ADMIN_APPROVED' | 'DRIVER_ASSIGNED' | 'IN_TRANSIT' | 'COMPLETED' | 'CANCELLED_BY_CONSIGNOR' | 'CANCELLED_BY_ADMIN' | 'CANCELLED_SYSTEM';
  goodType?: string;
  loadingDate?: string;
  loadingLocation?: string;
  offloadingLocation?: string;
  priceAmount?: number;
  priceCurrency?: string;
  publicId?: string;
  requiredVehicleNumber?: number;
  requiredVehicleType?: string;
}